third_party/WebKit/LayoutTests/fast/dom/focus-on-hidden-object.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<object id="hidden" style="visibility:hidden;" width="50" height="50"></object>
<script>
test(function(){
    var hidden = document.getElementById('hidden');
    hidden.focus();
    assert_not_equals(hidden, document.activeElement, 'hidden objects did not receive focus');
}, 'Hidden objects should not receive focus');
</script>
